% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abk_one.page.R
\name{abk_one.page}
\alias{abk_one.page}
\title{Abkhaz one page search}
\usage{
abk_one.page(my.url)
}
\arguments{
\item{my.url}{Vector with a link}
}
\description{
Post query to the Abkhaz Text Corpus (http://baltoslav.eu/apsua/index.php). Don't forget to check main page for advanced query language.
}
\examples{
ava_one.page.get("http://baltoslav.eu/apsua/index.php?xxx=\%D0\%B1\%D1\%8B\%D0\%B7\%D1\%88\%D3\%99\%D0\%B0&met=kon&ks=ca&mova=en#wierch")
}
\author{
George Moroz <agricolamz@gmail.com>
}

