% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ava_corpus.R
\name{ava_corpus}
\alias{ava_corpus}
\title{Avar search}
\usage{
ava_corpus(query, n_results = 20, kwic = TRUE, write = FALSE)
}
\arguments{
\item{query}{Vector of length 1 which contain query.}

\item{n_results}{Integer. Defines number of examples from the corpus. By default is 10}

\item{kwic}{Logical. KWIC (key word in context) is the format for resulted lines. If TRUE, then it returns a dataframe with query in the middle and left and right contexts. If FALSE, then it returns each result in one string. By default is TRUE.}

\item{write}{This argument writes a file in the working derictory (see function getwd() and setwd() for more information). If FALSE, then it creates a dataframe in Global Environment. Otherwise function writes a .tsv file with the name frome the argument value. By default is FALSE.}
}
\description{
Post query to the Avar Text Corpus (http://baltoslav.eu/avar/index.php). Don't forget to check main page for advanced query language.
}
\examples{
ava_corpus("магIарул")
ava_corpus("магIарул*")
ava_corpus("ва(ц|с)азе", n_results = 50)
ava_corpus("(чIужу|руччаби)", kwic = F)
}
\author{
George Moroz <agricolamz@gmail.com>
}

