% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ava_one.page.R
\name{ava_one.page}
\alias{ava_one.page}
\title{Avar one page search}
\usage{
ava_one.page(my.url)
}
\arguments{
\item{my.url}{Vector with a link}
}
\description{
Post query to the Avar Text Corpus (http://baltoslav.eu/avar/index.php). Don't forget to check main page for advanced query language.
}
\examples{
ava_one.page.get("http://baltoslav.eu/avar/index.php?xxx=магIарул*&met=kon&ks=ca&f=d&mova=en#wierch")

}
\author{
George Moroz <agricolamz@gmail.com>
}

