% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_coresp.R
\name{ql_coresp}
\alias{ql_coresp}
\title{Query-lines correspondencies}
\usage{
ql_coresp(query_word, lines)
}
\arguments{
\item{query_word}{Vector with a query words}

\item{lines}{Vector with a result lines}
}
\description{
Corpuses from http://baltoslav.eu/ can have two query tags in one result example. This function get query-lines correspondencies. It returns index of the query words for corect splitting of result lines into kwic parts.
}
\examples{
ql_coresp(c("a", "b", "c"), c("a b c", "c"))

}
\author{
George Moroz <agricolamz@gmail.com>
}

